#115 design: component name, feed css refactoring
